Check authoritative nameservers for a domain
DNS caching causes all sorts of issues, because checking the domain once caches the record
on many layers between the querying host and the authoritative nameservers.

To get around caching, simply ask the authoritative nameservers for the domain, bypassing
the DNS in the local network and all networks in between.