Bump version of public_suffix
Ruby 2.7 adds some new warnings. In particular, previously if a method
was defined with keyword arguments, you could create a hash and pass it to
the method, and Ruby wouldn't complain.

Now, however, Ruby wants you to explicitly pass **the_hash instead of just
the_hash where keyword arguments are expected.

The version of public_suffix that this codebase was relying on emitted the
following warnings:

    lib/ruby/gems/2.7.0/gems/public_suffix-3.1.1/lib/public_suffix/list.rb:51:
         warning: Using the last argument as keyword parameters is deprecated;
         maybe ** should be added to the call
    lib/public_suffix/list.rb:69:
         warning: The called method `parse' is defined here

v4.0.3 fixes these warnings.

I had to set the version specifier in the gemspec to >= 3.0 in order to
manage to upgrade github/github, otherwise bundle gets its dependencies
in a tizzy.

Once github/github is updated, we could update this to be ~> 4.0