Add spec to ensure that web requests are passed the correct proxy option
This test does not ensure that web requests are actually *made* with the
specified proxy. That is currently beyond the ability of WebMock to
assert (https://github.com/bblimke/webmock/issues/753). However Typhoeus
does raise an exception when an invalid option is passed to its Request
object, so we can be reasonably assured that Typhoeus will in fact
receive the intended proxy url and use it.