Fix `Checkable` incompatibility with `pp` gem
The `Checkable#pretty_print` method clobbers the mixin `Object#pretty_print`
added by the `pp` gem, which is used by IRB. This commit makes the method respond
to any number of arguments. This fixes an issue where IRB would raise an error
trying to pretty print an object when it was the return value of an expression.